client/requirements.txt
pydantic>=2.0
websocket-client>=1.6
